+++ Blogifier.Core/Data/Repositories/CustomRepository.cs
@@ -23,11 +23,11 @@ namespace Blogifier.Core.Data.Repositories
 
             if(parentId == 0)
             {
-                dbFields = _db.CustomFields.Where(f => f.CustomType == customType).OrderBy(f => f.Title);
+                dbFields = _db.CustomFields.Where(f => f.CustomType == (int)customType).OrderBy(f => f.Title);
             }
             else
             {
-                dbFields = _db.CustomFields.Where(f => f.CustomType == customType && f.ParentId == parentId).OrderBy(f => f.Title);
+                dbFields = _db.CustomFields.Where(f => f.CustomType == (int)customType && f.ParentId == parentId).OrderBy(f => f.Title);
             }
 
             return Task.Run(() => Load(dbFields));
@@ -35,26 +35,26 @@ namespace Blogifier.Core.Data.Repositories
 
         public Task<Dictionary<string, string>> GetBlogFields()
         {
-            var dbFields = _db.CustomFields.Where(f => f.CustomType == CustomType.Application && f.ParentId == 0).OrderBy(f => f.Title);
+            var dbFields = _db.CustomFields.Where(f => f.CustomType == (int)CustomType.Application && f.ParentId == 0).OrderBy(f => f.Title);
             return Task.Run(() => Load(dbFields));
         }
 
         public Task<Dictionary<string, string>> GetUserFields(int profileId)
         {
-            var dbFields = _db.CustomFields.Where(f => f.CustomType == CustomType.Profile && f.ParentId == profileId).OrderBy(f => f.Title);
+            var dbFields = _db.CustomFields.Where(f => f.CustomType == (int)CustomType.Profile && f.ParentId == profileId).OrderBy(f => f.Title);
             return Task.Run(() => Load(dbFields));
         }
 
         public string GetValue(CustomType customType, int parentId, string key)
         {
-            var field = _db.CustomFields.Where(f => f.CustomType == customType && f.ParentId == parentId && f.CustomKey == key).FirstOrDefault();
+            var field = _db.CustomFields.Where(f => f.CustomType == (int)customType && f.ParentId == parentId && f.CustomKey == key).FirstOrDefault();
             return field == null || field.CustomValue == null ? string.Empty : field.CustomValue;
         }
 
         public Task SetCustomField(CustomType customType, int parentId, string key, string value)
         {
             var dbField = _db.CustomFields
-                .Where(f => f.CustomType == customType && f.ParentId == parentId && f.CustomKey == key)
+                .Where(f => f.CustomType == (int)customType && f.ParentId == parentId && f.CustomKey == key)
                 .FirstOrDefault();
 
             if (dbField != null)
@@ -69,7 +69,7 @@ namespace Blogifier.Core.Data.Repositories
                     CustomKey = key,
                     CustomValue = value,
                     Title = key,
-                    CustomType = customType,
+                    CustomType = (int) customType,
                     ParentId = parentId,
                     LastUpdated = SystemClock.Now()
                 });
